Refactor package name and fix comments
