--- model/src/main/java/io/github/ygojson/model/atomic/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Contains the atomic models.
- * </br>
- * Atomic models are those where the information
- * is not related to other models and contain only
- * the information that represents.
- */
-package io.github.ygojson.model.atomic;
