+++ model/src/main/java/io/github/ygojson/model/data/definitions/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Contains the definitions for data models.
+ * </br>
+ * The definitions for the data models are values
+ * that are helping the description of the model.
+ * </br>
+ * On YGOJSON, we try to minimize the number of definitions
+ * to have as flat as possible data models.
+ */
+package io.github.ygojson.model.data.definitions;
