+++ model/src/main/java/io/github/ygojson/model/data/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains the Data Models.
+ * </br>
+ * Data Models are those that exists independently
+ * and only related between them by IDs.
+ */
+package io.github.ygojson.model.data;
