--- model/src/main/java/io/github/ygojson/model/enums/package-info.java
@@ -1,9 +1,0 @@
-/**
- * Contains the enums that are hardcoded into YGOJSON.
- * </br>
- * Enums are models that are defined within YGOJSON.
- * Other models, as they depend on releases (i.e., monster types)
- * by the game, are generated and thus are defined in the schemas
- * as strings.
- */
-package io.github.ygojson.model.enums;
