+++ model/src/test/java/io/github/ygojson/model/data/acceptance/PackageSettings.java
@@ -1,0 +1,9 @@
+package io.github.ygojson.model.data.acceptance;
+
+/**
+ * Settings for the approval tests configuration.
+ */
+public class PackageSettings {
+
+    public static String ApprovalBaseDirectory = "../resources";
+}
