+++ model/src/test/resources/io/github/ygojson/model/data/acceptance/SchemaValidationTest.testAtomicSchemas.Card.approved.txt
@@ -1,0 +1,96 @@
+{
+  "type" : "object",
+  "properties" : {
+    "id" : {
+      "type" : "string",
+      "required" : true,
+      "format" : "uuid"
+    },
+    "name" : {
+      "type" : "string"
+    },
+    "identifiers" : {
+      "type" : "object",
+      "properties" : {
+        "konamiId" : {
+          "type" : "integer"
+        },
+        "password" : {
+          "type" : "integer"
+        },
+        "passwordAlt" : {
+          "type" : "integer"
+        },
+        "yugipediaPageId" : {
+          "type" : "integer"
+        }
+      }
+    },
+    "cardType" : {
+      "type" : "string",
+      "enum" : [ "monster", "spell", "trap" ]
+    },
+    "spellType" : {
+      "type" : "string",
+      "pattern" : "[a-z]+"
+    },
+    "trapType" : {
+      "type" : "string",
+      "pattern" : "[a-z]+"
+    },
+    "effectText" : {
+      "type" : "string"
+    },
+    "monsterTypes" : {
+      "type" : "array",
+      "items" : {
+        "type" : "string"
+      }
+    },
+    "flavorText" : {
+      "type" : "string"
+    },
+    "attribute" : {
+      "type" : "string",
+      "pattern" : "[a-z]+"
+    },
+    "atk" : {
+      "type" : "integer"
+    },
+    "atkUndefined" : {
+      "type" : "boolean"
+    },
+    "def" : {
+      "type" : "integer"
+    },
+    "defUndefined" : {
+      "type" : "boolean"
+    },
+    "level" : {
+      "type" : "integer"
+    },
+    "pendulumEffect" : {
+      "type" : "string"
+    },
+    "pendulumScale" : {
+      "type" : "integer"
+    },
+    "materials" : {
+      "type" : "string"
+    },
+    "linkRating" : {
+      "type" : "integer"
+    },
+    "linkArrows" : {
+      "type" : "array",
+      "items" : {
+        "type" : "string",
+        "enum" : [ "top-left", "top-center", "top-right", "middle-left", "middle-right", "bottom-left", "bottom-center", "bottom-right" ]
+      },
+      "maxItems" : 8
+    },
+    "xyzRank" : {
+      "type" : "integer"
+    }
+  }
+}
